requirements.txt
Pillow==3.4.2
mock==2.0.0
